+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Debugger - Button to disconnect current session (stay listening)
+- [#2644] Debugger - Button to sort variables alphabetically
 
 ### Fixed
 - [#2643] APNG export - images containing multiple IDAT chunks
@@ -4174,6 +4175,7 @@ Major version of SWF to XML export changed to 2.
 [alpha 9]: https://github.com/jindrapetrik/jpexs-decompiler/compare/alpha8...alpha9
 [alpha 8]: https://github.com/jindrapetrik/jpexs-decompiler/compare/alpha7...alpha8
 [alpha 7]: https://github.com/jindrapetrik/jpexs-decompiler/releases/tag/alpha7
+[#2644]: https://www.free-decompiler.com/flash/issues/2644
 [#2643]: https://www.free-decompiler.com/flash/issues/2643
 [#2645]: https://www.free-decompiler.com/flash/issues/2645
 [#2639]: https://www.free-decompiler.com/flash/issues/2639